contexts: refactored to Isolation manager
